webadmin: persist grid columns hidden-by-default
Persist the list of hidden-by-default columns that were requested
by the user to be visible.

Usually columns are visible-by-default. User can hide them and this
request is persisted in local storage. Some columns however are marked
as hidden programmatically when initializing the grid. Such
hidden-by-default columns require a reversed approach - the request
to show them must be persisted to prevent hiding the column each time
the grid is created.

Fixes bug scenario:
1. show a column hidden by default
2. log out and log in again
3. check the column visibility
Expected result: column is visible

Change-Id: I1d63d1a36a5fa531a8bc760e1640d0518b210caf
Bug-Url: https://bugzilla.redhat.com/1752751
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>