core: skip unplugged disks when assigning addresses to scsi disks
Let's say that we have a VM with a single SCSI disk - this disk would
be assigned with unit=0 in its address.

Next, when trying to plug a second SCSI disk we will:
1. Iterate all the disk, in this case the both SCSI disks
2. See that the first disk is mapped to unit 0
3. See that the second disk is not assigned with an address
4. Map the second disk to the next available unit, e.g., 1
5. Later, find the next available unit for the second disk, e.g., 2

So the second disk will be assigned with unit=2 although it could be
assigned with unit 1.

This patch improves this algorithm by skipping the unplugged devices
while iterating the disks (1). That way in the example above, the
second disk will be assigned with unit=1.

Change-Id: Ib3442edeb8508376f77d4ce31d3ad94d72be4fd4
Bug-Url: https://bugzilla.redhat.com/1855305
Signed-off-by: Arik Hadas <ahadas@redhat.com>