core: support merge bitmaps in a cold merge operation
If a cold snapshot merge is done for a VM that contains backups,
the new 'merged' base volume will not contain the new bitmaps if a backup
was taken with the top volume. Also, if a bitmap exists on the top volume
and on the base volume it should be merged to the base volume bitmap.

The engine should instruct the host to merge the bitmaps to
the base volume. A new flag was added for that reason to the API verb.

The new merge_bitmaps flag will be set to 'true' only if the disk
participates in an existing checkpoint.

Change-Id: Ieb7f10ab677e2078a671e2a2d7c43b35c315f6e5
Bug-Url: https://bugzilla.redhat.com/1861673
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>