core: write host devices to OVF
This patch will allow writing host devices into the OVF on non-OVA
flows. This is relevant mostly for snapshots. Previously, when the
user had host devices connected to the VM, they were dropped once
previewed or commited. Now it will remain as long the dedicated host in
the OVF exists in the cluster.

Change-Id: I8e8c8cdcf8aab7d6b12d6eac7dc35a6bc60a7d3f
Bug-Url: https://bugzilla.redhat.com/1869317
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>