core: expose OVA's OVF
This patch will allow the engine to expose the VM's OVF as the OVA's OVF
output. The disks populated size is omitted and sould be generated by
the client. Under each disk element: `ovf:populatedSize=<sizeInBytes>`.
The disks will change to COW+Sparse, as the disk should always be
converted to it during the OVA creation.
The OVF representation will allow the creation of OVAs by external
scripts.

Change-Id: I361deec4424c7b95f6bab50ac15c086f013f0de3
Bug-Url: https://bugzilla.redhat.com/1876234
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>