core: set custom bios type for hosted-engine vm
This patch sets custom bios type for hosted engine VM to prevent its
chipset or firmware from changing when the cluster's bios type changes
in existing deployments (new hosted-engine deployments would set the
custom bios type of the hosted-engine VM to q35+seabios when the
hosted engine VM is created).

The reason is twofold:
1. When changing the cluster's firmware to UEFI/UEFI+SecureBoot,
the hosted-engine VM fails to boot
2. When changing the cluster's chipset, the hosted engine VM may fail
to start due to incorrect adjustments of the VM specification.

We have no easy way to tell what chipset the hosted engine VM runs
with but a good indication is the controller of its CD-ROM - in
x86_64 clusters (which is the case for hosted engine environments),
it can either be IDE (when the vm starts) on i440fx or SATA on q35.

This patch does the following: hosted engine VMs that are not set
with custom bios type already will be set with custom bios type that
is determined by the existence of an IDE controller - in case an IDE
controller exists, they'll be set with i440fx chipset and otherwise
with q35+seabios.

This would cover the following cases:
1. The common case: by default the hosted engine VM is set with
bios type = cluster-default and resides in a q35+seabios cluster so
this change would set its custom bios type to the existing one.

2. The reported "problematic" case in which the cluster's bios type
is changed to q35+UEFI/UEFI+SecureBoot: the hosted engine VM would
not be able to boot so we can assume that it has not been restarted
since making that change. Here, we will change the firmware back to
SeaBIOS (the OVF will be updated and the VM could boot again).

3. The uncommon case in which the cluster's bios type is set to
i440fx: if the VM has not been started then it won't have i440fx-
specific settings (IDE controller, PCI addresses, etc) and thus it
is safe to set the VM with custom bios type = q35+seabios.
Otherwise, we will find that the VM is set with an IDE controller,
which means that in order to change it to q35 we need to adjust its
settings and thus we set its custom bios type to i440fx instead.

4. The "hacky" case in which the VM was modified and set with
custom bios type = i440fx already: we set the custom bios type to
i440fx because changing the chipset to q35 might require changes to
the VM devices. We know of just a few such "hacky" cases.

Change-Id: I357d5687bc9799b17026260d700ce5c52fee9940
Bug-Url: https://bugzilla.redhat.com/1871694
Signed-off-by: Arik Hadas <ahadas@redhat.com>