webadmin: change USB Support field to checkbox
At the moment, there are just 2 UsbPolicy supported - DISABLED and
ENABLED_NATIVE. There is no need for a combobox anymore and we can use
checkbox in the VM popup dialog instead. This allows to unify the field
naming between popup dialog (USB Support) and general pages (USB Policy)
and use just USB (Enabled).

Note that this is just FE change only that does not require the change
of the backend model.

Change-Id: Iefed3ad9f7cb7354a711792b5ff04a02142b4372
Bug-Url: https://bugzilla.redhat.com/1862785
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>