search: enable to search with keyword prefix
Search engine enables to enter free text pattern in a search expression, this
kind of search looks for the pattern on all the entity fields.

In the case that the free text pattern starts with a keyword (column
name) that is related to the searched entity, search engine fails to
find the expected results and generates an error.

For example, if you have a cluster named "namedCluster" and you are
searching for "Cluster:name*" , you will get an empty result and an
error complaining on illegal search, also if you are doing that from UI ,
you will see that the character '*' is marked with red.

The reason for that is that when the search engine analyses the search
string and finds a reserved keyword (column name) it expects an
expression in the format "<keyword><op><pattern>" where <op> is one of
"<", ">", "=", "!=", "<=", ">="
So , when the parser finds "name" in our example, it looks for one of
the operators mentioned above and finds "*" , this is the reason why it
fails to get the correct result.

This patch adds another check when a keyword is found, it simply checks
that it really has the format "<keyword><op><pattern>" by examining the
character appearing after the keyword (if exists), if it is not a blank
or a known operator then the search is categorized and executed as a
free text search on values on the entity columns that has the keyword as
a prefix.

A new test was added as well to check that when a keyword prefix is used
in free text search it will retrieve the correct and expected results.

Change-Id: I14c5e90bcae7bd9a41c2e3e1b4fae9d7fe04233d
Bug-Url: https://bugzilla.redhat.com/1797717
Signed-off-by: Eli Mesika <emesika@redhat.com>