ansible: replace shell with command in vdsm deploy role
Ansible `shell` command shouldn't be used unless necesarry (e.g. when on
needs pipelines) and `command` command should be used instead. Replace
shell with command in vdsm deploy role.

Ansible lint requires command and shell commands to have change_when
value so that Ansible is able to decide if the taks changed state or
not. As vdsm-tool doesn't allow to determine if the configuration was
changed or not, we always report that the state has changed.

For some reason in case of Configure LVM filter task Ansible doesn't
complain about missing changed_when. So in this task changed_when is
skipped.

Change-Id: I623317ed5ef69bcbc791ab9302d6594dc956a5a3
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>