core: update volume format after move
When moving an a RAW/sparse image to a block storage domain the image is
converted to COW. This change however, is not persisted after LSM, and if
the VM is restarted the image will be passed on as RAW despite being
COW on the storage.

This patch updates the image in the database after the move is complete.

Change-Id: I6588fc4fb3a3eda3535b3d1d74a2999d97b303f6
Bug-Url: https://bugzilla.redhat.com/1877790
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>