core: support copy bitmaps in a cold storage migration operation
If a cold storage migration is done for a VM that contains backups,
the new 'copied' disk will not contain the backup bitmaps.

The engine should instruct the host to copy the bitmaps to
all the disk volumes.
A new flag was added for that reason to the copy_data API verb.

The new copy_bitmaps flag will be set to 'true' only if the disk
participates in an existing checkpoint.

Change-Id: I61a30343658d5e34b166a53127d77a22c6845901
Bug-Url: https://bugzilla.redhat.com/1861671
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>