core: prevent a failure of command by compesnation
This patch will prevent failure of import commands running with
CommandCompensationPhase.END_COMMAND. When the engine service starts the
compensation takes action and rollback the database. This can cause a
successful command that runs async to fail.

This patch will check the import commands the compensation wants to
execute on engine startup and skip it in case the child command still
active.

Change-Id: I2a9d7a07c105608fec680273267bbbba27d560d3
Bug-Url: https://bugzilla.redhat.com/1871792
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>