engine: fixing score=0 after HE redeploy
This patch provides a fix for an issue with he score being set to 0
after host being re-installed with HE_DEPLOY option.

Details:
The scenario is as the following:
1) put he host into maintenance
2) reinstall with  HE_UNDEPLOY & 'put to active'
3) when active put it back into maintenance
4) reinstall with HE_DEPLOY & 'put to active'

During activation (step 4) VDS in SetHaMaintenanceModeVDSCommand used
to be loaded during post construct phase when
'highlyAvailableIsConfigured' flag still being set to false
 because of steps (2,3).

SetHaMaintenanceModeVDSCommand must be invoked on hosts with HA
already configured. There is no need to check it inside the command as
the 'invoker code' should already be aware of it and this is the idea behind
the fix itself.

Change-Id: I5fe821ff8c251ea375ebbac01d40c00d03bc921a
Bug-Url: https://bugzilla.redhat.com/1859586
Signed-off-by: Artur Socha <asocha@redhat.com>