packaging: pki: Clean up qemu CA handling
We recently added support for using the pki scripts for managing also a
separate CA for qemu (for secure encrypted live migration). I later
noticed that "qemu-ca" appears in a few places where the engine ca was
expected. So, this patch cleans up relevant pki scripts and engine-setup
plugins. This includes fixing conf/template files but not
already-created certificates. Details:

* files names use cleanup

So far, we had constant names only for the .in files, but didn't have
_IN in the constants' names. We used string manipulation to get the
other files' names (the template and conf files generated from them),
which made it hard to quickly search the code to see what code handles a
particular file. Now, I am creating a constant for each relevant conf/
template file and only use this constant to refer to it, instead of the
string manipulation. I also add some simple data structures to correlate
the templates with the files generated from them.

* pki scripts

- They already had code for doing particular things based on the ca file
that was passed to them to work on. Clean this up to be common to all
scripts and for conf files.

- Support passing --ca-file also to pki-enroll-pkcs12.sh

This is currently not needed, and I am adding it mainly for debugging.

Can be useful in the future, if we want a single command to create
and sign a keypair on the engine machine using the qemu ca. Without
this patch, this requires creating the private key and request
using other means (e.g. calling openssl directly) and then calling
pki-enroll-request.sh.

- Fail pki-enroll-request.sh if conf is missing

pki-enroll-request.sh used to use cert.conf hard-coded. cert.conf was
created by pki-create-ca.sh, also hard-coded there. Now, conf is
parametric (either cert.conf or qemu-cert.conf). Generation is still the
same (by pki-create-ca.sh), also parametric. Since there is a slight
chance of a bug around this, I explicitly check for its existence and
abort, if missing, with:

    "${CERT_CONF} is missing, Cannot sign certificate"

* Fix pki uninstall

Do not add:

- /etc/pki/ovirt-engine/certs/$name
- /etc/pki/ovirt-engine/keys/$name

Do add:

- /etc/pki/ovirt-engine/keys/$name.p12

Fixes bz 1811593. Not really mandatory to be as part of this patch, but
makes sense.

* Fix _miscUpgrade

- Process templates for qemu-ca to their own files.

- Fix broken templates on update, if they were not manually changed.
Emit a suitable message in any case.

* Move stuff from _create_ca to engine (_create_primary_ca)

Things that are engine-specific, should not be in _create_ca which is
used also for qemu ca.

Bug-Url: https://bugzilla.redhat.com/1875386
Bug-Url: https://bugzilla.redhat.com/1811593
Change-Id: Icafc99a35545d9514540e3dd0084d3eedc5859ad
Signed-off-by: Yedidyah Bar David <didi@redhat.com>