pki: Use genpkey instead of genrsa
It seems that genrsa uses md5 in the process (if encrypting the key),
regardless of options passed, and this fails in FIPS mode.

openssl manpage says that genrsa is superseded by genpkey.

So: Use genpkey.

Also use AES256 instead of 3DES, as it's considered more secure
(although 3DES is still allowed in FIPS mode, it seems).

I am pushing this patch to fix the linked bug in FIPS mode.
For mere fixing, it's enough to not encrypt the private key at all, and
this is still not less secure - as it's a temporary file, later saved
inside an encrypted PKCS#12 file. But doing above (genpkey, AES) makes
more sense going-forward, and I do not see any compatibility or
performance issues that might arise.

Bug-Url: https://bugzilla.redhat.com/1875363
Change-Id: I84f807b2dfa48edb35939cbdd0472ed1fabcd76a
Signed-off-by: Yedidyah Bar David <didi@redhat.com>