core: Support shallow download
When ImageTransfer's shallow is true and using the NBD backend for a
qcow2 disk, configure vdsm NBD server to expose only the specified
volume, instead of the entire chain starting at the volume.

If shallow is not specified, we expose the entire chain when using the
NBD backend.

When using the file backend shallow has no effect. The file backend
always expose a single volume.

This change depends on ovirt-engine-api-model patch:
https://gerrit.ovirt.org/c/111312/

Change-Id: Ife7773810139e36b55f1406c668bda853037a20b
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>