core: Add option to list active disk snapshots
The API filters out the active disk snapshots by default, making it hard
to get the complete backing chain of a disk. The backing chain is
required for backing up the current content of the disk, keeping the
snapshots.

Since existing applications may depend on the current behavior, we
cannot fix this issue by including the current snapshot by default. Add
an "include_active" query parameter to enable the new behavior.

To include the active disks snapshots, user can use:

    GET /storagedomains/sd-id/disksnapshots?include_active

Now that the active disk snapshot is accessible, is it possible to
download all snapshots in a consistent manner, using:

    types.ImageTransfer(
        snapshot=types.DiskSnapshot(id=disk_snapshot_id)
        ...

Previously to download the active disk snapshot you had to use:

    types.ImageTransfer(
        disk=types.Disk(id=disk_id)
        ...

Change-Id: I154dc106437f8ea48c81cb182ec6623b82de7265
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>