core: Expose disk snapshot parent
The parent is needed to understand disk snapshot order, and to
reconstruct disk snapshots when backing up snapshots.

Internally we use a special empty Guid when an image does not have a
parent. In the API it is nicer to report no value in this case.

This change depends on api model change:
https://gerrit.ovirt.org/c/111479

Change-Id: Id0963d089de661fa5b30551d7554a67b0c5e2f88
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>