core: support hosted engine devices on first boot
This patch will let Hosted Engine VM to correlate the devices when no
alias exists. This is relevant to the deployment part of that VM. Which
may result in unplugging a device and later on a duplication of the same
device. This will also keep showing an error in the engine log about
the device.

Change-Id: I6f4eec24e549f27adaba19cbf493adfb71c45879
Bug-Url: https://bugzilla.redhat.com/1881471
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>