core: Forbid Disk.reduce() for active image
Disk.reduce() validation is incorrect, allowing reduce of the active
image used by a running VM. This is unsafe since the disk allocation may
increase while we reduce the logical volume.

I think this bug exists since code was added in:

commit c58e21d8fa2dcf4510add765289a244ce915bf13
Author: Daniel Erez <derez@redhat.com>
Date:   Wed May 9 17:11:20 2018 +0300

    core: expose ReduceImageCommand to clients

Fix the validation to forbid this operation on the active disk image.

When trying to reduce a disk used by running VM via the SDK, the
operation fails now with:

    ovirtsdk4.Error: Fault reason is "Operation Failed". Fault detail is
    "[action type failed reduce image not supported for active image]".
    HTTP response code is 400.

Change-Id: I79c9b37d1401364c99d002850c3b41a2767f0404
Bug-Url: https://bugzilla.redhat.com/1887202
Related-To: https://bugzilla.redhat.com/1443963
Signed-off-by: Nir Soffer <nsoffer@redhat.com>