core: update cpu config for Skylake cpus
Update cpu configuration for secure Skylake-Client and Skylake-Server
for 4.4 and 4.5 versions.

Secure variants of the mentioned CPUs now use -noTSX-IBRS qemu CPU
version. This version does not require hle and rtm flags to be present
on the host system and enables spec-ctrl by default.

Note that if the host system has disabled tsx (hle and rtm flags are not
present) it is not possible to use the default CPU variant because qemu
does not have Skylake-Client-noTSX CPU version.

Change-Id: I78bcd946177b43b34e76dd9a5a4cf726fa4a9835
Bug-Url: https://bugzilla.redhat.com/1881325
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>