core: add 'backup_mode' column to base_disk table
Add 'backup_mode' for each disk which is part of the backup.
The new 'backup_mode' will indicate if the backup that will be taken
for the disk will be 'full' or 'incremental' backup.
If the disk doesn't participate in any backup then the backup mode
will be 'null'.

The 'backup_mode' must be kept in the database for later use by the
user, after the backup was taken the user will use this to understand
which backup was taken for each disk and download it accordingly
using imageio.

Change-Id: Ia7f62d10774d118a5185d01f05e72fbef2d6c6ac
Bug-Url: https://bugzilla.redhat.com/1861674
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>