Add Support for UUID of oVirt Instance
To be able to verify the uniqueness of a particular installation
of oVirt (for data protection purposes), each oVirt installation will
from now on have an 'instanceId' (UUID) generated for it. The ID is
generated upon installation (or upgrade, for setups where it doesn't yet
exist) during the execution of the database scripts. It s stored in
vdc_options table (with 'InstanceId' as key) and returned to clients
for the request:

GET .../ovirt-engine/api

<api>
   ...
   <product_info>
      <instance_id>XXX</instance_id>
   /<product_info>
</api>

Change-Id: I5459afd7e1c53ed5c50f30077a90c38718b53ae1
Bug-Url: https://bugzilla.redhat.com/1825020
Signed-off-by: Ori_Liel <oliel@redhat.com>