Remove Dependency on Api-Explorer
The API-Explorer (ovirt-engine-api-explorer) presents some javascript
vulnerabilities, and on the other hand has not been in use for a while.
Thus, it has been decided to remove it from oVirt.

This patch marks it as obsolete in the .spec file, which will result
in this package's removal upon oVirt version upgrade.

Change-Id: Ib904ac53455888a3715a2952f781a5f18357dec2
Signed-off-by: Ori_Liel <oliel@redhat.com>