core: Disable monitor CPU flag in nested AMD VMs
Due to libvirt bug https://bugzilla.redhat.com/1798004, VMs on AMD
EPYC models may not start in a nested host.  The problem is that
libvirt doesn't handle missing `monitor' CPU flag properly.

It can be remedied by adding <cpu> feature manually:

  <feature name="monitor" policy="disable"/>

This will be safe also once libvirt is fixed.

This patch adds the flag, based on some guesswork.  We assume the flag
is needed if:

- The host CPU is some sort of AMD EPYC; and
- the host CPU is missing `monitor' flag.

Change-Id: I80891190de3e08ade26220f80e632cbef71e3806
Bug-Url: https://bugzilla.redhat.com/1689362
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>