core: Assume miimon=100 as default for bond mode 4
The bond option "miimon=100" is now set by default on bond
mode 4 by Linux kernel. Because vdsm does not report default
bond options, this option is not reported by vdsm anymore.
If Engine tries to detect if the bond options are the
Engine's default ones, Engine's reference should be what is
expected to be reported by vdsm.

Change-Id: Ic0d3802d8c18aa98212d9560b98b242dbcef93f9
Bug-Url: https://bugzilla.redhat.com/1890071
Signed-off-by: Dominik Holler <dholler@redhat.com>