core: reduce number of queries of vds dynamic on create vm
The main reason to use MemoizingSupplier when building the domain XML is to
reduce the number of queries from the database. However, it appears that we
use more and more fields from VdsDynamic now and for every field we query
VdsDynamic from the database due to how we use MemoizingSuppliers.

This patch reduces the number of queries of VdsDynamic by LibvirtVmXmlBuilder
from five to one.

Change-Id: I453714609e4f506630c742e48187c2dd1fcf0998
Signed-off-by: Arik Hadas <ahadas@redhat.com>