core: allow cold VM incremental backup
Cold VM backup is now supported for a powered-off VM.

Creating a cold VM backup will create a new bitmap for each
disk that participates in the backup. During the backup operation,
the checkpoint database entity will not contain the checkpoint XML
because this backup is done without libvirt intervention.

The creation of the missing checkpoint XML while the engine
redefines the VM's checkpoints will be done in the following patches.

Stopping a cold VM backup doesn't require any VDSM operation.

Change-Id: I242369e68cd14753b1c2519c0725bd11ccc98d86
Bug-Url: https://bugzilla.redhat.com/1891470
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>