utils: fix cpu to numa pinning
The calculation to split the number of CPUs into each NUMA node was
incorrect in case we had more than two nodes. In that case we hit out of
bound to the CPUs exist in the NUMA node. Now, in that case we will
assign it to the maximal CPUs in that NUMA node.

Change-Id: I474ad282b355240691409719730d1b462e6bfbb6
Bug-Url: https://bugzilla.redhat.com/1891293
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>