core: initiate down vms handling before restarting vms
When rebooting VMs (with cold-reboot) or when highly-available VMs stop
running unexpectedly, the engine automatically attempts to restart them
upon detecting the VMs in Down state (with exit-reason != Normal) or not
reported at all.

This patch initiates the handling of Down/unreported VMs (for instance,
applying next-run configuration, rollback quota, etc) before initiating
the aforementioned restarts - this does not guarantee that the former is
executed before the latter but it increases the chances that next-run
configurations would be applied before trying to restart VMs that are
rebooted via cold-reboot.

Change-Id: I14658c5234781eedd528028366b36ffe2f6ce6c0
Bug-Url: https://bugzilla.redhat.com/1880251
Signed-off-by: Arik Hadas <ahadas@redhat.com>