core: prevent potential state of a running vm with run_on_vds = null
When we call warm-reboot (i.e., reboot from within the guest while the qemu
process keeps running) and receive a successful response, we change the VM
state to RebootInProgress. This is typically ok since it takes time to
interact with the guest and for the guest to shut down, but theoretically
this can lead to a scenario in which the VM is set with run_on_vds = null
and its status = RebootInProgress - a state we must avoid.

This patch prevents this by introducing RebootVmVDSCommand that locks the
VM during the call to the 'reboot' operation and changing the VM status so
the monitoring would not be able to change the VM simultaneously.

Change-Id: I5311719aca9ac5be5b803a8548ebcf0a97a67698
Bug-Url: https://bugzilla.redhat.com/1880251
Signed-off-by: Arik Hadas <ahadas@redhat.com>