scheduling: fix NullPointerException on CPU usage
When the host is switching to non-operational and then coming back to Up
state and on the same time we run a VM. It is possible we have a race on
the VDS status. In that case, the host will appear in the scheduling
hosts as optional but some of its value might set with null.

It is required to check the possibility to null value within the
CpuOverloadPolicyUnit filter, dropping the host from the list.

Change-Id: Iae33e871a5eb8b131bce777602278dfa5e57b371
Bug-Url: https://bugzilla.redhat.com/1889394
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>