Fix timezone warning when daylight offset
The offset can't just be parsed from the timezone string.
This as it doesn't take daylight offset into account.
Use the .getOffset() call to have correct offset.

To do this, the check of whether the offset that is reported
from the guest matches the expected offset of the timezone
at the given time is moved to the backend.

Change-Id: Ied6d5d3269005be1bfd1257cab81296d0d757423
Bug-Url: https://bugzilla.redhat.com/1881026
Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>
Signed-off-by: Arik Hadas <ahadas@redhat.com>