core: increase size of leaf when copying disks
When copying disks with copy_data the target leaf will be very small as
it only takes the required size which is based on the amount of written
data. This will make the VM pause for extension very quickly.

This patch adds an additional 1GB or increases the size of the leaf up
to the disk's size to avoid an immediate pause.

Change-Id: I5a8adc19e477d1b7ca69e7af0caca5ebda8a7d9d
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>