core: Warn when not restoring stateless
There is a warning when checking for the latest version
when bringing down a pool VM during validation of
the version when restoring a stateless VM which does
not have the new template version set.

Added checking to ensure that the use latest version is also
set before printing the warning and before removing the Virtual
Machine from the database.

Change-Id: I9d80d559239cd025dddcd7d048d26e0fb7bd223d
Bug-Url: https://bugzilla.redhat.com/1888142
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>