core: Use the network provider of the cluster for hosts
Keeping the network provider of a host in sync with the
default network provider of the cluster might result in
inconsistencies.
Taking into account that it is not possible anymore to
configure the network provider per host, the commands should
use the network provider of the cluster, instead the network
provider of the host.

Change-Id: I0f625cf7f73e37d5e9fd4602c7985885d29793c6
Bug-Url: https://bugzilla.redhat.com/1885132
Signed-off-by: Dominik Holler <dholler@redhat.com>