core: prevent status=RebootInProgress and run_on_vds=null
We must avoid a scenario in which a VM is set with status != Down/Suspended
and run_on_vds = null. If we reach this state, The VM would be stuck with
its current status (unless the database is changed manually).

Due to a combination of:
1. Getting an event quickly when the VM switches to Down state; and
2. VMs remain in powering up state for 60 sec even when installed with
qemu-guest-agent (which is fixed already)

We noticed that the following leads to the above mentioned state that
should be prevented:

1. VM is starting
2. Next-run configuration is set for the VM
3. Reboot is initated before the VM reaches up state

As we reach #3 when the VM is set with next-run configuration, the 'reboot'
operation calls 'shutdown'. As the VM is still in Powering up state in #3,
the 'shutdown' command calls 'destroy'.

VDSM destroys the VM and triggers an event. The even typically arrives to
the engine before the engine manages to change the VM status to
RebootInProgress and so as a result of processing the event, the VM is
changed to: status = Down + run_on_vds = null and later on, we change the
VM to RebootInProgress while its run_on_vds field is already null.

To prevent this, we are going to lock the VM in a way that would prevent
the VMs monitoring from processing events/polling that may change any of
the VM properties while the 'reboot' operation is executed.

Change-Id: If56ad696eea09d770781c45d74e88de1d2192a52
Bug-Url: https://bugzilla.redhat.com/1880251
Signed-off-by: Arik Hadas <ahadas@redhat.com>