core: clone VmInit when creating pool VMs
AddVmPool/UpdateVmPool may trigger multiple AddVm commands and specify the same
instance of VmInit in the parameters of all those commands. Each instance of
VmInit is placed in VmBase and its id changes when it is retrieved using
VmBase#getVmInit to the VM's id.

The problem with the above is that it leads to the following race:
1. First instance of AddVm is executed, setting the id within VmInit to X.
2. The first instance of AddVm reaches its asynchronous phase.
3. Second instance of AddVm is executed, setting the id within VmInit to Y.
4. The id within the VmInit is changed back to X by the infrastructure.
5. The second instance of AddVm tries to persist VmInit with id=X.
-> We end up trying to persist VmInit with id that already exists and thus fail.

This patch fixes this by cloning VmInit when VmBase is cloned. That way, each
instance of AddVm is provided with its own instance of VmInit and therefore one
instance of AddVm cannot affect another instance of AddVm and so each AddVm
command will persist VmInit with id that equals to that of the VM being added.

Change-Id: I287d3ec9887aca2511d98f76f448d8babaeebcfe
Bug-Url: https://bugzilla.redhat.com/1758216
Signed-off-by: Arik Hadas <ahadas@redhat.com>