core: Remove Host devices when removing a Host
When removing a Host, the Virtual Machine's Host
Device was not being removed from the Virtual
Machines' Device database table nor was the Host
removed from the Virtual Machines' Dedicated Host
List.

Now when a Host is removed from the system, the Host is
also removed from the Virtual Machines' Dedicated Host List
and the Virtual Machines' Host Device entry is also removed
for the Host in question.

Change-Id: Ifcad55e417be3e2a533e1d5d06cc39c2764c852c
Bug-Url: https://bugzilla.redhat.com/1886750
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>