vdsbroker: set serial to lun disks
This patch will allow setting the serial for non-passthrough direct LUN
disks to the VM. This will later be used by VDSM to map the disk with
its guest logical name. It is required to do so, since the serial will
be missing or made up, and we won't be able to map it without it.

Change-Id: I9c4c52627eb5c5b575a5a9a35ff940e9be22b93e
Bug-Url: https://bugzilla.redhat.com/1859092
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>