core: CloneVm: Use passed disk parameters
In the new CloneVmCommand implementation, diskInfoDestinationMap is
built from the VM snapshot that is created for cloning. As a
consequence, this map does not include modifications (target storage
domain, disk alias etc.) that are passed in diskInfoDestinationMap in
the command parameters.

This patch adds merging of diskInfoDestinationMap passed in the command
parameters into diskInfoDestinationMap created by the command.

Change-Id: Ia3627e9a1687251d5b7391816c85fccc8a75dfb6
Bug-Url: https://bugzilla.redhat.com/1895697
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>