core: Snapshot disk remains locked if image transfer times out
In case that download snapshot disks using the SDK is
failed, the image status isn't changed back to OK and
the diskImage remains in a locked state.

It is happening since for snapshot disks the
ImageGroupId parameter not set in case of image
download action.
This patch removes the check if the ImageGroupId is
NULL and sets the image status to be OK again in
case of a failure.

Change-Id: I1db05f025d339c000a17f990747e3ddacb209088
Bug-Url: https://bugzilla.redhat.com/1796231
Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>