core: preserve bios-type setting on live-merge
With the current implementation of bios-type, when we read a VM from OVF
we store the original cluster's bios type in the 'clusterBiosTypeOrigin'
field and when writing a VM to an OVF we expect the 'clusterBiosType' to
store the cluster's bios-type.

So previously, when we initiated live-merge on a VM that has no custom
bios-type but that uses the cluster's bios-type it lead to:
1. Reading the VM from the OVF of the modified snapshot, thus the
cluster's bios-type was stored using the 'clusterBiosTypeOrigin' field.
2. Writing the modified configuration of that VM to the snapshot - but
the 'clusterBiosType' field was not set, thus the bios-type was not
written to the OVF.

Consequently, the snapshot that the data is merged to contained the VM
configuration that is not set with the original bios-type - but with an
empty bios-type which is interpreted as i440fx+seabios.

This patch copies the 'clusterBiosTypeOrigin' field to the
'clusterBiosType' field in this flow. The true fix for this though
should be to stop pointing to the cluster's bios type in the VM
configuration (which is a work in progress).

Change-Id: I231c3a04a763830dede88cd44307734141dc582c
Bug-Url: https://bugzilla.redhat.com/1899768
Signed-off-by: Arik Hadas <ahadas@redhat.com>