core: allow requiring consistency for VM backup
With the new require_consistency flag, the user/backup-app will have
the option to choose for each backup if this backup should fail if
the VM failed to freeze before starting the backup
(which may lead to inconsistent data in the backup) or not.

In order to require consistency for backup, the following request should
be used:

POST /ovirt-engine/api/vms/123/backups?require_consistency

<backup>
  <from_checkpoint_id>previous-checkpoint-uuid</from_checkpoint_id>
  <disks>
      <disk id="disk-uuid" />
      ...
  </disks>
</backup>

The default value for require_consistency is false.

Change-Id: I0b40856ea08da0938946a79676eab9498c6ae22f
Bug-Url: https://bugzilla.redhat.com/1894413
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>