webadmin: Set numa tune mode per node
Allow to set the tune mode per numa node.
This is done in Numa Pinning dialog when you
expand the numa node details.

Note that the setting of the numa tune mode
for all nodes was removed from the VM popup dialog.
It is now not possible to select one value that would
be applied to all nodes. It was removed to prevent
the confusion between settings on the VM level (VM popup
dialog) and numa node level (Numa Pinning dialog).

Change-Id: I4bdac4ad5d209efecc27ef03a9905ce1de951918
Bug-Url: https://bugzilla.redhat.com/1729897
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>