core: add per virtual numa tune
This feature will allow setting the NUMA tune mode per each vNode that
the VM has. The value is settable once the VM has vNode but in use only
when there is pinning of vNode to pNode.

As part of the change, the NUMA tune mode removed from the VM entity.
The API and backend is fully supported within this patch. UI is
semi-supported and will be fully supported on upcoming patch.

There is no need to have backward compatibility since we never saved the
VMs NUMA node tune to OVFs. A VM was set by default to `interleave` mode
with this patch it will act the same for vNodes without a specific tune.

Change-Id: Id50627fd1d17576e8dcc016e47d47a756eb5d2e3
Bug-Url: https://bugzilla.redhat.com/1729897
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>