core: monitor image removal tasks on update vm-version
When UpdateVmVersions tries to update a VM to the latest version of the
template it is based on, it removes the VM and re-create it on top of
the latest version of the template.

Previously, async tasks that were created during the removal of the VM
were not monitored by the UpdateVmVersion command. Now, UpdateVmVersion
would monitor those tasks and call AddVm upon their completion.

Change-Id: I6ba57e6eea7fefcaf239ddf7f52b5f1afc510dcf
Signed-off-by: Arik Hadas <ahadas@redhat.com>