core: Modify NICs after cloning a VM
If the clone VM modal contains changes to VM NICs, this changes need to
be executed after the cloning process is finished. Previously the
frontend called Add/Update/RemoveVmInterface commands just after
CloneVmCommand has returned. But now the cloning process starts by
creating a VM snapshot, so at the moment CloneVmCommand returns the
cloned VM still does not exist.

To solve this problem, VmInterfacesModifyCommand was added. It is
executed as the last phase of CloneVmCommand. The list of NICs is passed
to it through CloneVmParameters.

Change-Id: I5cbe7fccde10b5adaeb753eef9019f6ee2bd3e43
Bug-Url: https://bugzilla.redhat.com/1895695
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>